Refactoring models.rs and updating references throughout
