+++ src/models/mod.rs
@@ -1,0 +1,6 @@
+//! Contains all models used in `fetters`.
+
+pub mod job;
+pub mod sprint;
+pub mod status;
+pub mod title;
